+++ tests/code/conftest.py
@@ -1,11 +1,11 @@
 import pytest
 
-from madewithml.data import CustomPreprocessor
+from ai_ml_ops.data import CustomPreprocessor
 
 
 @pytest.fixture
 def dataset_loc():
-    return "https://raw.githubusercontent.com/GokuMohandas/Made-With-ML/main/datasets/dataset.csv"
+    return "datasets/dataset.csv"
 
 
 @pytest.fixture